backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25
redis==5.0.1
python-dotenv==1.0.0
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
tiktoken==0.5.2
chromadb==0.4.22
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3
